add: ConsultaDTO. fix:Datos de archivos de Consulta para exponer datos segun haga falta

+++ backend/src/main/java/com/medilatam/backend/Interface/IDoctorService.java
@@ -21,9 +21,9 @@ public interface IDoctorService {
     public void saveDoctor (Doctor doctor);
     
     //Eliminar un Objeto que sea Doctor
-    public void deleteDoctor (Integer id);
+    public void deleteDoctor (Long id);
     
     //Buscar un Objeto del tipo Doctor por Id
-    public Doctor findDoctor (Integer id);
+    public Doctor findDoctor (Long id);
     
 }
